feat: optimization catalog controller
Currently, The catalog controller and catalog table are run with an auto-incrementing id. I think it
would be better to use the id generated by spring's snowflake model algorithm. As well as querying
and deleting related functions, they are operated by the catalog id instead of using the catalog
name.

re #70